The newest version of project 1
It includes audio core and audio source file now. It has been tested.

+++ project/software/libraries/SDcard.h
@@ -7,7 +7,7 @@
 
 #ifndef SDCARD_H_
 #define SDCARD_H_
-#include <../Module1/Definitions.h>
+#include <../audio_core_test/Definitions.h>
 
 
 void sdcard_Init(void);
@@ -16,16 +16,16 @@ int sdcard_isFAT16(void);
 short int sdcard_FirstFile(char *directory, char *file_name);
 short int sdcard_NextFile(char *file_name);
 int sdcard_FileSize(short int file_handle);
-void sdcard_ReadFile(int read_data [], short int file_handle);
+void sdcard_ReadFile(short int read_data [], short int file_handle);
 short int sdcard_ReadByte(short int file_handle);
-void sdcard_WriteFile(int write_data [], short int file_handle);
+void sdcard_WriteFile(short int write_data [], short int file_handle);
 void sdcard_ListFiles(char *directory);
 
 short int sdcard_fopen(char *name, bool create_new_if_not_exist);
 void sdcard_fclose(short int file_handle);
 short int sdcard_getFattributes(short int file_handle);
 
-void printArray(int a[]);
+void printArray(short int a[]);
 typedef struct
 {
 	char *file_name;
